seperate out layerItem
