+++ web/src/beta/features/Editor/tabs/map/LeftPanel/Layers/index.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback, useState } from "react";
 
-import TextInput from "@reearth/beta/components/fields/TextInput";
 import Icon from "@reearth/beta/components/Icon";
 import * as Popover from "@reearth/beta/components/Popover";
 import PopoverMenuContent from "@reearth/beta/components/PopoverMenuContent";
-import Text from "@reearth/beta/components/Text";
 import type { LayerNameUpdateProps } from "@reearth/beta/features/Editor/useLayers";
 import type { NLSLayer } from "@reearth/services/api/layersApi/utils";
 import { styled } from "@reearth/services/theme";
+
+import LayerItem from "./LayerItem";
 
 type LayersProps = {
   layers: NLSLayer[];
@@ -74,87 +74,6 @@ const Layers: React.FC<LayersProps> = ({
   );
 };
 
-type LayerItemProps = {
-  id: string;
-  layerTitle: string;
-  onDelete: () => void;
-  onSelect: () => void;
-  onLayerNameUpdate: (inp: LayerNameUpdateProps) => void;
-};
-
-const LayerItem = ({ id, layerTitle, onDelete, onSelect, onLayerNameUpdate }: LayerItemProps) => {
-  const [isMenuOpen, setMenuOpen] = useState(false);
-  const toggleMenu = useCallback(() => {
-    setMenuOpen(prev => !prev);
-  }, []);
-  const [isEditing, setIsEditing] = useState(false);
-  const [editedTitle, setEditedTitle] = useState(layerTitle);
-
-  const handleDoubleClick = useCallback(() => {
-    setIsEditing(true);
-  }, []);
-
-  const handleTitleSubmit = useCallback(() => {
-    if (!editedTitle.trim()) {
-      return;
-    }
-    setIsEditing(false);
-    onLayerNameUpdate({ layerId: id, name: editedTitle });
-  }, [editedTitle, id, onLayerNameUpdate]);
-
-  const handleKeyUp = useCallback(
-    (e: { key: string }) => {
-      if (["Enter", "Return"].includes(e.key)) {
-        handleTitleSubmit();
-      }
-    },
-    [handleTitleSubmit],
-  );
-
-  const handleBlur = useCallback(() => {
-    setIsEditing(false);
-    handleTitleSubmit();
-  }, [handleTitleSubmit]);
-
-  return (
-    <ListItemContainer onClick={onSelect}>
-      <div onDoubleClick={handleDoubleClick}>
-        {isEditing ? (
-          <TextInput
-            value={editedTitle}
-            onChange={(newTitle: string) => setEditedTitle(newTitle)}
-            onKeyUp={handleKeyUp}
-            onBlur={handleBlur}
-          />
-        ) : (
-          <Text size="body">{layerTitle}</Text>
-        )}
-      </div>
-      <div>
-        <Popover.Provider open={isMenuOpen} onOpenChange={toggleMenu} placement="left-start">
-          <Popover.Trigger asChild>
-            <MenuIcon onClick={toggleMenu}>
-              <Icon icon="actionbutton" />
-            </MenuIcon>
-          </Popover.Trigger>
-          <Popover.Content>
-            <PopoverMenuContent
-              size="md"
-              items={[
-                {
-                  name: "Delete",
-                  icon: "bin",
-                  onClick: onDelete,
-                },
-              ]}
-            />
-          </Popover.Content>
-        </Popover.Provider>
-      </div>
-    </ListItemContainer>
-  );
-};
-
 const LayerContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -164,17 +83,6 @@ const AddLayerIcon = styled.div`
   padding: 2px;
   margin-bottom: 2px;
   align-self: flex-end;
-  cursor: pointer;
-`;
-
-const ListItemContainer = styled.div`
-  padding: 8px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const MenuIcon = styled.div`
   cursor: pointer;
 `;
